fix(coevolution): resolve integration, actor state, and database issues in parametric runs - Fix stale CSR array references in `AgentProcessor` during coevolution.   Added the `UpdateTopology` message and made `neighborsRefs`,   `neighborsWeights`, and `neighborBiases` reassignable (`var`) inside   `AgentProcessor`. `Network` now broadcasts updated array references to   all child actors after each rewiring iteration, preventing   `ArrayIndexOutOfBoundsException`. - Initialize `RoundRouter` database-saver actors in `ParametricOrchestrator`   during `preStart()` to avoid `NullPointerException` when executing runs   independently of the main `Monitor` actor. - Redirect `insertAgentsBatch` in `DatabaseManager` to query the legacy   database (`getConnectionLegacy`) rather than the active database,   resolving the `relation "public.agents" does not exist` error. - Use `.name` instead of `.toString` for `SilenceStrategy` and `SilenceEffect`   during agent batch writes to ensure compliance with PostgreSQL enums. - Update `DatabaseManager.getConsensusCount` and `getAvgRounds` to fetch   from `promueva_legacy.networks` when `GlobalState.APP_MODE.usesLegacyDB`   is active. - Fix snowflake integer overflows by altering `run_id` columns to `bigint`   in the legacy schema (`legacy_schema.sql`). - Fix line-endings (CRLF to LF) in database entrypoint scripts to prevent   initialization failures in the PostgreSQL Docker environment. - Add `ParametricTestRunner` to automate scaling tests across base-2 sizes   and write results to `parametric_results.csv`, handling shutdown timeouts.

+++ db/init/legacy_schema.sql
@@ -146,7 +146,7 @@ CREATE TABLE IF NOT EXISTS public.networks
     id                 uuid                          NOT NULL,
     run_time           bigint  DEFAULT '-1'::integer NOT NULL,
     build_time         bigint  DEFAULT '-1'::integer NOT NULL,
-    run_id             integer                       NOT NULL,
+    run_id             bigint                        NOT NULL,
     number_of_agents   integer                       NOT NULL,
     final_round        integer DEFAULT '-1'::integer NOT NULL,
     name               character varying(16)         NOT NULL,
@@ -186,7 +186,7 @@ CREATE TABLE IF NOT EXISTS public.agent_states_speaking
 
 CREATE TABLE IF NOT EXISTS public.generated_run_parameters
 (
-    run_id              integer NOT NULL,
+    run_id              bigint NOT NULL,
     degree_distribution real    NOT NULL,
     density             integer NOT NULL
 );
